+++ __tests__/cli/team.test.js
@@ -1,0 +1,6 @@
+import { runCLI } from '../utils'
+
+test('missing subcommand', async () => {
+  const out = await runCLI({ args: 'team', testForError: true })
+  expect(out).toMatchSnapshot()
+})